updating variations
