+++ blocks/init/src/blocks/assets/scripts/application-blocks-editor.js
@@ -13,7 +13,7 @@
  
  */
 
-import { registerBlocks } from '@eightshift/frontend-libs/scripts/editor';
+import { registerBlocks, registerVariations } from '@eightshift/frontend-libs/scripts/editor';
 import { Wrapper } from './../../wrapper/wrapper';
 import WrapperManifest from './../../wrapper/manifest.json';
 import globalSettings from './../../manifest.json';
@@ -23,12 +23,19 @@ registerBlocks(
   globalSettings,
   Wrapper,
   WrapperManifest,
-  require.context('./../../custom', true, /manifest\.json$/),
+  require.context('./../../custom', true, /manifest.json$/),
   require.context('./../../custom', true, /-block.js$/),
   require.context('./../../custom', true, /-hooks.js$/),
-  require.context('./../../custom', true, /-variations.js$/),
   require.context('./../../custom', true, /-transforms.js$/),
   require.context('./../../custom', true, /-icons.js$/),
+);
+
+registerVariations(
+  globalSettings,
+  WrapperManifest,
+  require.context('./../../variations', true, /manifest.json$/),
+  require.context('./../../variations', true, /-transforms.js$/),
+  require.context('./../../variations', true, /-icons.js$/),
 );
 
 // Run Wrapper hooks.
