+++ blocks/init/src/blocks/variations/button-block/manifest.json
@@ -1,0 +1,14 @@
+{
+  "parentName": "button",
+  "blockName": "button-full-width",
+  "title": "Button Full Width",
+  "description" : "Button Block Variation for full width style.",
+  "attributes": {
+    "button": {
+      "styleSizeWidth": "block"
+    }
+  },
+  "scope": [
+    "inserter"
+  ]
+}
