+++ scripts/editor/index.js
@@ -1,6 +1,5 @@
 export { getActions } from './get-actions';
 export { getPaletteColors } from './get-palette-colors';
-export { registerBlock } from './register-block';
-export { registerBlocks } from './register-blocks';
+export { registerBlocks, registerVariations } from './register-blocks';
 export { ucfirst } from './ucfirst';
 export { icons } from './icons';
